requirements.txt
azure-functions
azure-ai-formrecognizer>=4.2.0
azure-storage-blob>=12.14.1